feat: migrate auth fn to separate file
