+++ server/src/http/routes/etablissementRecruteur.controller.ts
@@ -32,8 +32,7 @@ import { validationOrganisation } from "../../services/bal.service.js"
 import { IUserRecruteur } from "../../common/model/schema/userRecruteur/userRecruteur.types.js"
 import { IRecruiter } from "../../common/model/schema/recruiter/recruiter.types.js"
 import { updateUserValidationHistory, getUser, createUser, updateUser, getUserValidationState, registerUser } from "../../services/userRecruteur.service.js"
-import { IAdresseV3 } from "../../common/model/schema/_shared/shared.types.js"
-import { authMiddleware } from "../../auth/passport-strategy.js"
+import authMiddleware from "../middlewares/authMiddleware.js"
 
 const getCfaRomeSchema = joi.object({
   latitude: joi.number().required(),
